feat: move clock.System to systemclock.New
Makes the externalclock and systemclock provide similar APIs for
different clock implementations.
