+++ systemclock/doc.go
@@ -1,0 +1,2 @@
+// Package systemclock provides a clock implementation for the system clock.
+package systemclock
